Added Windows Store build support

+++ engine/include/purpl/util/misc.h
@@ -57,12 +57,12 @@ extern PURPL_API u64 purpl_random(u64 limit);
 /// The current process ID
 ///
 /// \return Returns the current process ID
-extern PURPL_API u32 purpl_get_pid(void);
+extern PURPL_API u64 purpl_get_pid(void);
 
 /// The current thread ID
 ///
 /// \return Returns the current thread ID
-extern PURPL_API u32 purpl_get_tid(void);
+extern PURPL_API u64 purpl_get_tid(void);
 
 /// Trigger a debug breakpoint (debug builds only)
 extern PURPL_API void purpl_debug_break(void);
